[feature/tabar-setup]: resolve comment
