+++ app/src/main/java/com/sun/nhom5/screen/broadcast/BroadcastFragment.kt
@@ -1,0 +1,19 @@
+package com.sun.nhom5.screen.broadcast
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.sun.nhom5.R
+
+class BroadcastFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_broadcast, container, false)
+    }
+}
