+++ app/src/main/java/com/sun/nhom5/screen/example/SampleActivity.kt
@@ -1,21 +1,16 @@
-package com.example.android_template.example
+package com.sun.nhom5.screen.example
 
-import android.view.LayoutInflater
-import android.view.ViewGroup
 import android.widget.Toast
-import com.sun.nhom5.base.BaseFragment
-import com.sun.nhom5.databinding.FragmentSampleBinding
-import com.sun.nhom5.example.ExampleContract
-import com.sun.nhom5.example.ExamplePresenter
-import com.sun.nhom5.example.ITeamRepository
-import com.sun.nhom5.example.TeamRepository
+import com.sun.nhom5.databinding.ActivitySampleBinding
+import com.sun.nhom5.screen.base.BaseActivity
 
-class SampleFragment : BaseFragment<FragmentSampleBinding, ExamplePresenter>(), ExampleContract.View {
+class SampleActivity : BaseActivity<ActivitySampleBinding, ExamplePresenter>(),
+    ExampleContract.View {
 
     private lateinit var repository: ITeamRepository
 
-    override fun getViewBinding(inflater: LayoutInflater, container: ViewGroup?): FragmentSampleBinding {
-        return FragmentSampleBinding.inflate(inflater, container, false)
+    override fun getViewBinding(): ActivitySampleBinding {
+        return ActivitySampleBinding.inflate(layoutInflater)
     }
 
     override fun initializePresenter() {
@@ -36,15 +31,15 @@ class SampleFragment : BaseFragment<FragmentSampleBinding, ExamplePresenter>(), 
 
     override fun showMyTeams(teams: List<String>) {
         Toast.makeText(
-            context,
-            teams.joinToString(", "),
-            Toast.LENGTH_LONG
+            this,
+            teams.joinToString { it },
+            Toast.LENGTH_SHORT
         ).show()
     }
 
     override fun showLoading() {
         Toast.makeText(
-            context,
+            this,
             "Loading teams...",
             Toast.LENGTH_SHORT
         ).show()
@@ -55,7 +50,7 @@ class SampleFragment : BaseFragment<FragmentSampleBinding, ExamplePresenter>(), 
 
     override fun showError(message: String) {
         Toast.makeText(
-            context,
+            this,
             "Error: $message",
             Toast.LENGTH_LONG
         ).show()
