+++ app/src/main/java/com/sun/nhom5/screen/example/SampleFragment.kt
@@ -1,15 +1,18 @@
-package com.sun.nhom5.example
+package com.example.android_template.example
 
+import android.view.LayoutInflater
+import android.view.ViewGroup
 import android.widget.Toast
-import com.sun.nhom5.base.BaseActivity
-import com.sun.nhom5.databinding.ActivitySampleBinding
+import com.sun.nhom5.databinding.FragmentSampleBinding
+import com.sun.nhom5.screen.base.BaseFragment
+import com.sun.nhom5.screen.example.ExampleContract
+import com.sun.nhom5.screen.example.ExamplePresenter
+import com.sun.nhom5.screen.example.TeamRepository
 
-class SampleActivity : BaseActivity<ActivitySampleBinding, ExamplePresenter>(), ExampleContract.View {
+class SampleFragment : BaseFragment<FragmentSampleBinding, ExamplePresenter>(), ExampleContract.View {
 
-    private lateinit var repository: ITeamRepository
-
-    override fun getViewBinding(): ActivitySampleBinding {
-        return ActivitySampleBinding.inflate(layoutInflater)
+    override fun getViewBinding(inflater: LayoutInflater, container: ViewGroup?): FragmentSampleBinding {
+        return FragmentSampleBinding.inflate(inflater, container, false)
     }
 
     override fun initializePresenter() {
@@ -17,8 +20,6 @@ class SampleActivity : BaseActivity<ActivitySampleBinding, ExamplePresenter>(), 
     }
 
     override fun setupViews() {
-        repository = TeamRepository()
-        presenter = ExamplePresenter(repository)
         presenter?.attachView(this)
     }
 
@@ -30,15 +31,15 @@ class SampleActivity : BaseActivity<ActivitySampleBinding, ExamplePresenter>(), 
 
     override fun showMyTeams(teams: List<String>) {
         Toast.makeText(
-            this,
-            teams.joinToString { it },
-            Toast.LENGTH_SHORT
+            context,
+            teams.joinToString(", "),
+            Toast.LENGTH_LONG
         ).show()
     }
 
     override fun showLoading() {
         Toast.makeText(
-            this,
+            context,
             "Loading teams...",
             Toast.LENGTH_SHORT
         ).show()
@@ -49,7 +50,7 @@ class SampleActivity : BaseActivity<ActivitySampleBinding, ExamplePresenter>(), 
 
     override fun showError(message: String) {
         Toast.makeText(
-            this,
+            context,
             "Error: $message",
             Toast.LENGTH_LONG
         ).show()
