+++ app/src/main/java/com/sun/nhom5/screen/music/MusicFragment.kt
@@ -1,0 +1,19 @@
+package com.sun.nhom5.screen.music
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.sun.nhom5.R
+
+class MusicFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_music, container, false)
+    }
+}
