+++ app/src/main/res/layout/fragment_profile.xml
@@ -4,23 +4,18 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    android:orientation="vertical"
-    tools:context=".presentation.profile.ProfileFragment">
+    android:orientation="vertical">
 
     <Button
         android:id="@+id/btnSampleMVP"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:text="Navigate to Sample MVP" />
-        />
 
     <Button
         android:id="@+id/btnSampleAPI"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:text="Navigate to Sample API" />
-
-
-
 
 </LinearLayout>
