+++ app/src/main/res/layout/fragment_sample.xml
@@ -4,8 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".example.SampleActivity">
+    android:layout_height="match_parent">
 
     <Button
         android:id="@+id/loadMyTeamsButton"
