+++ app/src/main/res/menu/bottom_nav_menu.xml
