+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,25 +7,25 @@
 
     <fragment
         android:id="@+id/home_fragment"
-        android:name="com.example.android_template.presentation.home.HomeFragment"
+        android:name="com.sun.nhom5.screen.home.HomeFragment"
         android:label="Home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/broadcast_fragment"
-        android:name="com.example.android_template.presentation.broadcast.BroadcastFragment"
+        android:name="com.sun.nhom5.screen.broadcast.BroadcastFragment"
         android:label="Broadcast"
         tools:layout="@layout/fragment_broadcast" />
 
     <fragment
         android:id="@+id/music_fragment"
-        android:name="com.example.android_template.presentation.music.MusicFragment"
+        android:name="com.sun.nhom5.screen.music.MusicFragment"
         android:label="Music"
         tools:layout="@layout/fragment_music" />
 
     <fragment
         android:id="@+id/profile_fragment"
-        android:name="com.example.android_template.presentation.profile.ProfileFragment"
+        android:name="com.sun.nhom5.screen.profile.ProfileFragment"
         android:label="Profile"
         tools:layout="@layout/fragment_profile" >
         <action
